[#14/feature] Fix docker-compose script commands and fix instructions in the readme

+++ quasar.sh
@@ -2,15 +2,17 @@
 
 if [[ ! -d /var/www/app/node_modules ]]; then
   echo "~> installing dependencies"
-  if yarn install 2>/dev/null; then
+  if [[ which yarn ]]; then
      yarn install
   else
-     npm install
+    if [[ which npm ]]; then
+       npm install
+    fi
   fi
 fi
 
-if [[ ! -f /home/node/bin/node ]]; then
-  echo "~> export bin"
+if [[ ! -f /home/node/bin/node && -f /usr/local/bin/node ]]; then
+  echo "~> expose bin"
   cp /usr/local/bin/node /home/node/bin/node
   echo "~> fix permissions"
   chown -R node:node .
@@ -19,4 +21,4 @@ fi
 echo "Details: '$(pwd)' | '$(quasar -v)'"
 
 echo "~> starting dev"
-quasar $(cat /var/www/app/quasar.sh.env)
+quasar $(cat /var/www/app/quasar.env)
